Minor restructure, update readme
