+++ hood/crypto_trading/schema/__init__.py
